refactor(agent-manager): add max-lines lint rule and extract large files
Add ESLint max-lines error at 3000 lines for .ts/.tsx files in the
extension. Extract NewWorktreeDialog into its own file and move review/
file-tree CSS into agent-manager-review.css to bring both violating
files under the limit.

- AgentManagerApp.tsx: 3144 -> 2536 lines
- agent-manager.css: 3159 -> 2728 lines
